# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  authorization.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource_type:
        type: string
      updated_at:
        type: string
    type: object
  authorization.ResourcePermission:
    properties:
      action:
        description: Action type (e.g., "create", "read", "update", "delete")
        type: string
      created_at:
        type: string
      default_scope:
        description: Default permission scope (e.g., "own", "team", "all")
        type: string
      id:
        type: integer
      permission_id:
        description: 'Optional: legacy permission Id'
        type: integer
      resource_id:
        description: 'Optional: specific resource Id if applicable'
        type: string
      resource_type:
        description: Resource type (e.g., "project", "employee", etc.)
        type: string
      role_id:
        description: 'Optional: role Id for role-based permissions'
        type: string
      updated_at:
        type: string
      user_id:
        description: 'Optional: specific user Id if applicable'
        type: integer
    type: object
  authorization.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      permission_count:
        description: New field
        type: integer
      updated_at:
        type: string
    type: object
...
  types.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: info@base.al
    name: Base Team
    url: https://base.al
  description: This is the API documentation for Base Framework
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://base.al/terms
  title: Base Framework API
  version: 2.0.0
paths:
  /api/authorization/permissions:
    get:
      consumes:
      - application/json
      description: Get all permissions in the system
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/authorization.Permission'
                type: array
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all permissions
      tags:
      - Core/Authorization
...
  /authorization/check:
    post:
      consumes:
      - application/json
      description: Checks if a user has permission to perform an action on a resource
      parameters:
      - description: Permission check request
        in: body
        name: checkRequest
        required: true
        schema:
          properties:
            action:
              type: string
            organization_id:
              type: string
            resource_id:
              type: string
            resource_type:
              type: string
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Permission check result
          schema:
            properties:
              has_permission:
                type: boolean
            type: object
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Check user permission
      tags:
      - Core/Authorization
  /authorization/resource-permissions:
    post:
      consumes:
      - application/json
      description: Creates a resource-specific permission override
      parameters:
      - description: Resource permission to create
        in: body
        name: resourcePermission
        required: true
        schema:
          $ref: '#/definitions/authorization.ResourcePermission'
      produces:
      - application/json
      responses:
        "201":
          description: Resource permission created successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/authorization.ResourcePermission'
            type: object
        "400":
          description: Invalid resource permission data
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create resource permission
      tags:
      - Core/Authorization
  /authorization/resource-permissions/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a resource-specific permission override
      parameters:
      - description: Resource Permission Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Resource permission deleted successfully
          schema:
            properties:
              success:
                type: boolean
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete resource permission
      tags:
      - Core/Authorization
  /authorization/roles:
    get:
      consumes:
      - application/json
      description: Get all roles in the system
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/authorization.Role'
                type: array
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all roles
      tags:
      - Core/Authorization
    post:
      consumes:
      - application/json
      description: Creates a new role with the provided information
      parameters:
      - description: Role object to be created
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/authorization.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/authorization.Role'
            type: object
        "400":
          description: Invalid role data
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create a new role
      tags:
      - Core/Authorization
  /authorization/roles/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a role by its Id
      parameters:
      - description: Role Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role deleted successfully
          schema:
            properties:
              success:
                type: boolean
            type: object
        "403":
          description: System role cannot be deleted
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete a role
      tags:
      - Core/Authorization
    get:
      consumes:
      - application/json
      description: Retrieves a specific role by its Id
      parameters:
      - description: Role Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            properties:
              data:
                $ref: '#/definitions/authorization.Role'
            type: object
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get role by Id
      tags:
      - Core/Authorization
    put:
      consumes:
      - application/json
      description: Updates an existing role with the provided information
      parameters:
      - description: Role Id
        in: path
        name: id
        required: true
        type: string
      - description: Updated role object
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/authorization.Role'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/authorization.Role'
            type: object
        "400":
          description: Invalid role data
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "403":
          description: System role cannot be modified
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update a role
      tags:
      - Core/Authorization
...
  /authorization/roles/{id}/permissions/{permissionId}:
    delete:
      consumes:
      - application/json
      description: Removes a permission from a role
      parameters:
      - description: Role Id
        in: path
        name: id
        required: true
        type: string
      - description: Permission Id
        in: path
        name: permissionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Permission revoked successfully
          schema:
            properties:
              success:
                type: boolean
            type: object
        "404":
          description: Role or permission not found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Revoke permission from role
      tags:
      - Core/Authorization
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key for authentication
    in: header
    name: X-Api-Key
    type: apiKey
  BearerAuth:
    description: Enter your token with the prefix "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"